+++ src/GafferModule/NodeBinding.cpp
@@ -60,7 +60,7 @@ namespace
 
 struct UnaryPlugSlotCaller
 {
-	boost::signals::detail::unusable operator()( boost::python::object slot, PlugPtr p )
+	void operator()( boost::python::object slot, PlugPtr p )
 	{
 		try
 		{
@@ -70,14 +70,13 @@ struct UnaryPlugSlotCaller
 		{
 			PyErr_PrintEx( 0 ); // clears the error status
 		}
-		return boost::signals::detail::unusable();
 	}
 };
 
 struct BinaryPlugSlotCaller
 {
 
-	boost::signals::detail::unusable operator()( boost::python::object slot, PlugPtr p1, PlugPtr p2 )
+	void operator()( boost::python::object slot, PlugPtr p1, PlugPtr p2 )
 	{
 		try
 		{
@@ -87,13 +86,12 @@ struct BinaryPlugSlotCaller
 		{
 			PyErr_PrintEx( 0 ); // clears the error status
 		}
-		return boost::signals::detail::unusable();
 	}
 };
 
 struct ErrorSlotCaller
 {
-	boost::signals::detail::unusable operator()( boost::python::object slot, const Plug *plug, const Plug *source, const std::string &error )
+	void operator()( boost::python::object slot, const Plug *plug, const Plug *source, const std::string &error )
 	{
 		try
 		{
@@ -103,7 +101,6 @@ struct ErrorSlotCaller
 		{
 			ExceptionAlgo::translatePythonException();
 		}
-		return boost::signals::detail::unusable();
 	}
 };